fix(build): déplace types/constantes alternants hors de page.tsx (types.ts)
Un fichier page Next.js ne peut exporter que le composant + la config de route.
Le split alternants exportait des interfaces/constantes depuis page.tsx -> build cassé.
Déplacé dans alternants/types.ts ; page + _components importent depuis ./types.
